chore: move remove message to conv messages view model
